refactor: extract tcp_info handling to sockstat module
Move socket statistics code from HTTP engine to separate module for better modularity and reusability.
